implementations/centralized-discovery-service/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
aiohttp
dnspython
pydantic
python-dotenv
alembic
spacy
pytest
pytest-asyncio
pydantic-settings
httpx